int = float. AI killer

+++ src/main/java/io/github/nottoxicdev/BasicEnemy.java
@@ -8,7 +8,7 @@ import java.awt.Rectangle;
 public class BasicEnemy extends GameObject {
     private Handler handler;
 
-    public BasicEnemy(int x, int y, ID id, GroupID gid, Handler handler) {
+    public BasicEnemy(Float x, Float y, ID id, GroupID gid, Handler handler) {
         super(x, y, id, gid);
 
         this.handler = handler;
@@ -38,11 +38,11 @@ public class BasicEnemy extends GameObject {
             g2d.draw(getBounds());
         }
         g.setColor(Color.MAGENTA);
-        g.fillRect(x, y, 24, 24);
+        g.fillRect((int) x, (int) y, 24, 24);
     }
 
     public Rectangle getBounds() {
-        return new Rectangle(x, y, 24, 24);
+        return new Rectangle((int) x, (int) y, 24, 24);
     }
 
 }
